feat: create user with post request & refactor: in memory database
