+++ src/resources/in-memory-db.js
@@ -1,0 +1,8 @@
+const users = [
+  { id: '1', name: 'Max', login: 'trshkv', password: '12345' },
+  { id: '2', name: 'Sanya', login: 'agent007', password: '12345' },
+  { id: '3', name: 'Maryan', login: 'soviet-citizen', password: '12345' },
+  { id: '4', name: 'kate', login: 'soviet-citizen', password: '12345' },
+];
+
+module.exports = { users };
