+++ src/resources/users/user.model.js
@@ -1,11 +1,11 @@
-const uuid = require('uuid');
+const { v4: uuidv4 } = require('uuid');
 
 class User {
   constructor({
-    id = uuid(),
+    id = uuidv4(),
     name = 'USER',
     login = 'user',
-    password = 'P@55w0rd'
+    password = 'P@55w0rd',
   } = {}) {
     this.id = id;
     this.name = name;
@@ -16,6 +16,10 @@ class User {
   static toResponse(user) {
     const { id, name, login } = user;
     return { id, name, login };
+  }
+
+  static fromRequest(body) {
+    return new User(body);
   }
 }
