+++ src/resources/users/user.service.js
@@ -1,6 +1,12 @@
 const usersRepo = require('./user.memory.repository');
 
 const getAll = () => usersRepo.getAll();
+
 const getUserByID = (userID) => usersRepo.getUserByID(userID);
 
-module.exports = { getAll, getUserByID };
+const create = (user) => {
+  usersRepo.create(user);
+  return user;
+};
+
+module.exports = { getAll, getUserByID, create };
